Feat: Adds data field to AppError
Extends the `AppError` to include an optional `data` field.

This allows for passing additional information along with the error,
which can be useful for debugging and providing more context to the client.

The error handler is updated to include the `data` field in the response,
both in development and production environments.